+++ ugbase/lib_grid/algorithms/remeshing/simple_grid.h
@@ -51,7 +51,7 @@ void PrintSimpleGrid(SimpleGrid& sg);
  */
 template <class TPosAcc, class TIntAcc, class TNormAcc>
 bool ObtainSimpleGrid(SimpleGrid& sgOut, Grid& grid,
-						VertexBase* vrt1, VertexBase* vrt2, int size,
+						VertexBase* vrt1, VertexBase* vrt2, size_t size,
 						TPosAcc& aaPos, TNormAcc& aaNorm,
 						TIntAcc& aaInt);
